Only split BenchSpec
